Resolves blank code values
The code in 'templates/cat1/_2.16.840.1.113883.10.20.24.3.23.cat1.erb'
generates a blank value for 'participant/participantRole/code' is

Ex:
  <code code="" codeSystem=""/>

Should be:
  <code nullFlavor="UNK"/>

The if statement present in the template tests whether the code hash is
present, when it seems that it should instead test whether the value in
"code['code']" is present (e.g. not nil).

Doing this allows CAT1 tests in Cypress to pass, correctly placing:
  <code nullFlavor="UNK"/>

...into the CAT1 files when the "code['code']" value is nil, rather than
inserting an empty string.